refactor all tests running DBT to use a common test base class
